move Relese Status enum to constants module
