+++ constants/src/main/java/org/jboss/pnc/enums/ReleaseStatus.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rest.restmodel.causeway;
+package org.jboss.pnc.enums;
 
 import org.jboss.pnc.enums.MilestoneReleaseStatus;
